+++ src/Util/ClientFactoryAbstract.php
@@ -15,7 +15,7 @@ use Paysera\Component\RestClientCommon\Middleware\Authentication\MacAuthenticati
 use Paysera\Component\RestClientCommon\Middleware\Authentication\OAuthAuthentication;
 use Paysera\Component\RestClientCommon\Client\ApiClient;
 use Paysera\Component\RestClientCommon\Middleware\Exception\RequestExceptionMiddleware;
-use Paysera\Component\RestClientCommon\Middleware\GuzzleMiddlewareInterface;
+use Paysera\Component\RestClientCommon\Middleware\GuzzleMiddlewareProviderInterface;
 
 /**
  * @api
@@ -25,8 +25,8 @@ abstract class ClientFactoryAbstract
     const DEFAULT_BASE_URL = '';
     const AUTH_BASE_URL = 'https://wallet.paysera.com/oauth/v1/';
 
-    /** @var GuzzleMiddlewareInterface[] */
-    private $middlewares = [];
+    /** @var GuzzleMiddlewareProviderInterface[] */
+    private $middlewareProviders = [];
 
     private static $availableAuthTypes = [
         BasicAuthentication::TYPE,
@@ -47,9 +47,9 @@ abstract class ClientFactoryAbstract
         return new static($options);
     }
 
-    public function addMiddleware(GuzzleMiddlewareInterface $middleware): void
+    public function addMiddlewareProvider(GuzzleMiddlewareProviderInterface $middlewareProvider): void
     {
-        $this->middlewares[] = $middleware;
+        $this->middlewareProviders[] = $middlewareProvider;
     }
 
     public function createApiClient(array $options)
@@ -149,8 +149,8 @@ abstract class ClientFactoryAbstract
 
         $stack->unshift((new RequestExceptionMiddleware())->getMiddlewareFunction());
 
-        foreach ($this->middlewares as $middleware) {
-            $stack->push($middleware->getMiddlewareFunction());
+        foreach ($this->middlewareProviders as $middlewareProvider) {
+            $stack->push($middlewareProvider->getMiddleware());
         }
 
         return $client;