+++ CHANGELOG.md
@@ -6,8 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## 2.8.0
 ### Added
-- `GuzzleMiddlewareInterface` for defining Guzzle middleware components
-- `ClientFactoryAbstract::addMiddleware()` for injecting middleware into client factories
+- `GuzzleMiddlewareProviderInterface` for defining Guzzle middleware components
+- `ClientFactoryAbstract::addMiddlewareProvider()` for injecting middleware into client factories
 
 ## 2.7.2
 ### Fixed